models/pyproject.toml
[tool.poetry]
name = "models"
version = "0.1.0"
description = ""
authors = ["Vid Kersic <vid.kersic@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
giza-osiris = "^0.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
